+++ src/Braintree/IPlanGateway.cs
@@ -9,5 +9,11 @@ namespace Braintree
     {
         List<Plan> All();
         Task<List<Plan>> AllAsync();
+        Result<Plan> Create(PlanRequest request);
+        Task<Result<Plan>> CreateAsync(PlanRequest request);
+        Plan Find(string id);
+        Task<Plan> FindAsync(string id);
+        Result<Plan> Update(string id, PlanRequest request);
+        Task<Result<Plan>> UpdateAsync(string id, PlanRequest request);
     }
 }